+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add `Registration::setFrontEndUser()` and cache it (#508)
 
 ### Changed
-- Move more tests to the new testing framework (#507, #509)
+- Merge the testing registration model into the regular model (#510)
+- Move more tests to the new testing framework (#507, #509, #510)
 - Change the casing of `Registration::setFrontEndUserUID` (#506)
 - Allow empty user data for registrations (#505)
 - Improve the code autoformatting (#502)